LayoutTests/imported/w3c/web-platform-tests/css/css-cascade/scope-deep-expected.txt


FAIL Deep @scope nesting assert_equals: expected "rgb(0, 128, 0)" but got "rgb(0, 0, 0)"
